+++ api/routes/index.js
@@ -1,10 +1,10 @@
 const express = require("express");
 const userRouter = require("./userRouter");
-const mainRouter = require("./mainRouter");
+const cartRouter = require("./cartRouter");
 
 const router = express.Router();
 
 router.use("/users", userRouter);
-router.use("/", mainRouter);
+router.use("/carts", cartRouter);
 
 module.exports = router;
